pyproject.toml
[project]
name = "a-python-project-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "accelerate>=1.10.1",
    "ddgs>=9.8.0",
    "duckduckgo-search>=8.1.1",
    "easyocr>=1.7.2",
    "fastapi>=0.121.2",
    "googlesearch-python>=1.3.0",
    "gradio>=4.44.1",
    "ipykernel>=6.31.0",
    "pillow>=10.4.0",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.2",
    "timm>=1.0.22",
    "typer>=0.19.2",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest-cov>=7.0.0",
    "mkdocs>=1.5",
    "mkdocs-material>=9.0",
]
